lib: lte_lc: Extract and refactor helper functions
Helper functions, mostly for parsing, in the link controller are
abstraced and moved to a separate file.
This makes them easier to maintain and test.
