+++ lib/lte_link_control/CMakeLists.txt
@@ -5,4 +5,5 @@
 #
 
 zephyr_library()
-zephyr_library_sources_ifdef(CONFIG_LTE_LINK_CONTROL lte_lc.c)
+zephyr_library_sources(lte_lc.c)
+zephyr_library_sources(lte_lc_helpers.c)
